Api status code, validacao e tela

+++ app/Http/Controllers/API/ConsultaController.php
@@ -10,16 +10,20 @@ use Session;
 
 class ConsultaController extends BaseController
 {
+    protected $cpf;
+
     /**
      * Instantiate a new UserController instance.
      */
-    public function __construct()
+    public function __construct(Cpf $cpf)
     {
         session_start();
         if(!isset($_SESSION['start_date']))
         {
             $_SESSION['start_date'] = date('Y-m-d H:i:s');
         }
+
+        $this->cpf = $cpf;
     }
 
     /**
@@ -30,23 +34,46 @@ class ConsultaController extends BaseController
      */
     public function cpf(Request $request)
     {
-        $input = $request->all();
-        
-        $cpf = preg_replace('/\D/','',$input['cpf']);
-
-        $consulta_cpf = Cpf::where('cpf', '=', $cpf)->firstOrFail();
-
         try {
 
-            if (is_null($consulta_cpf)) 
+            $input = $request->all();
+            
+            $input['cpf'] = preg_replace('/\D/','',$input['cpf']);
+
+            
+            $valida_cpf = $this->cpf->validaCPF($input['cpf']);
+
+            if($valida_cpf)
             {
-                return $this->sendError('consulta_cpf not found.');
+
+                $consulta_cpf = Cpf::where('cpf', '=', $input['cpf'])->firstOrFail();
+
+                if (is_null($consulta_cpf)) 
+                {
+                    return $this->sendError('cpf não encontrado.')->setStatusCode(404);
+                }
+
+                $consulta_cpf->count = ($consulta_cpf->count + 1);
+                $consulta_cpf->save();
+
+                if(!isset($_SESSION['consultas_counter']))
+                {
+                    $_SESSION['consultas_counter'] = 0;
+                }
+
+                if(count($_GET) > 0)
+                {
+                    $_SESSION['consultas_counter']++;
+                }
+                
+                return $this->sendResponse($consulta_cpf->toArray(), 'consulta_cpf retrieved successfully.')->setStatusCode(200);
             }
 
-            return $this->sendResponse($consulta_cpf->toArray(), 'consulta_cpf retrieved successfully.');
+            return $this->sendError('cpf invalido.')->setStatusCode(404);
+
 
         } catch (Exception $e) {
-            return $this->sendResponse($input['cpf'], $e->getMessage); 
+            return $this->sendResponse($input['cpf'], $e->getMessage)->setStatusCode(400);; 
         }
     }
 
@@ -58,33 +85,114 @@ class ConsultaController extends BaseController
      */
     public function store(Request $request)
     {
-        $input = $request->all();
+        try {
+            $input = $request->all();
 
-        $input['cpf'] = preg_replace('/\D/','',$input['cpf']);
+            $input['cpf'] = preg_replace('/\D/','',$input['cpf']);
 
-        $validator = Validator::make($input, [
-            'cpf' => 'required'
-        ]);
+            $validator = Validator::make($input, [
+                'cpf' => 'required'
+            ]);
 
-        if($validator->fails()){
-            return $this->sendError('Validation Error.', $validator->errors());       
+            if($validator->fails()){
+                return $this->sendError('Validation Error.', $validator->errors())->setStatusCode(400);       
+            }
+
+            $valida_cpf = $this->cpf->validaCPF($input['cpf']);
+
+            if($valida_cpf)
+            {
+                $cpf_new = Cpf::firstOrNew($input);
+                $cpf_new->save();
+            
+                return $this->sendResponse($cpf_new->toArray(), 'Cpf criado com sucesso.')->setStatusCode(200);
+            }
+
+            return $this->sendError('cpf invalido.')->setStatusCode(404);
+
+        } catch (Exception $e) {
+            return $this->sendResponse($input['cpf'], $e->getMessage)->setStatusCode(400);; 
         }
+    }
 
-        $cpf = Cpf::firstOrNew($input);
-        //$cpf->count = ($cpf->count + 1);
-        $cpf->save();
-        
-        if(!isset($_SESSION['postCounter']))
-        {
-            $_SESSION['postCounter'] = 0;
+    /**
+     * Store a newly created resource in storage.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @return \Illuminate\Http\Response
+     */
+    public function block(Request $request)
+    {
+        try{
+
+            $input = $request->all();
+
+            $input['cpf'] = preg_replace('/\D/','',$input['cpf']);
+
+            $validator = Validator::make($input, [
+                'cpf' => 'required'
+            ]);
+
+            if($validator->fails()){
+                return $this->sendError('Validation Error.', $validator->errors())->setStatusCode(400);       
+            }
+
+            $valida_cpf = $this->cpf->validaCPF($input['cpf']);
+
+            if($valida_cpf)
+            {
+                $cpf_update =  Cpf::where('cpf', '=', $input['cpf'])->firstOrFail();   
+                $cpf_update->blocked = 1;
+                $cpf_update->save();
+            
+                return $this->sendResponse($cpf_update->toArray(), 'Cpf bloqueado com sucesso.')->setStatusCode(200);
+            }
+
+            return $this->sendError('cpf invalido.')->setStatusCode(404);
+
+        } catch (Exception $e) {
+            return $this->sendResponse($input['cpf'], $e->getMessage)->setStatusCode(400);; 
         }
+    }
 
-        if(count($_POST) > 0)
-        {
-            $_SESSION['postCounter']++;
+    /**
+     * Store a newly created resource in storage.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @return \Illuminate\Http\Response
+     */
+    public function unblock(Request $request)
+    {
+        try{
+
+            $input = $request->all();
+
+            $input['cpf'] = preg_replace('/\D/','',$input['cpf']);
+
+            $validator = Validator::make($input, [
+                'cpf' => 'required'
+            ]);
+
+            if($validator->fails()){
+                return $this->sendError('Validation Error.', $validator->errors())->setStatusCode(400);       
+            }
+
+            $valida_cpf = $this->cpf->validaCPF($input['cpf']);
+
+            if($valida_cpf)
+            {
+                $cpf_update =  Cpf::where('cpf', '=', $input['cpf'])->firstOrFail();   
+                $cpf_update->blocked = 0;
+                $cpf_update->save();
+            
+                return $this->sendResponse($cpf_update->toArray(), 'Cpf desbloqueado com sucesso.')->setStatusCode(200);
+            }
+
+            return $this->sendError('cpf invalido.')->setStatusCode(404);
+
+        } catch (Exception $e) {
+            return $this->sendResponse($input['cpf'], $e->getMessage)->setStatusCode(400);; 
         }
-
-        return $this->sendResponse($cpf->toArray(), 'Cpf criado com sucesso.');
     }
 
     /**
@@ -95,13 +203,25 @@ class ConsultaController extends BaseController
      */
     public function destroy(Request $request)
     {
-        $input = $request->all();
+        try{
+            $input = $request->all();
 
-        $input['cpf'] = preg_replace('/\D/','',$input['cpf']);
+            $input['cpf'] = preg_replace('/\D/','',$input['cpf']);
+                
+            $valida_cpf = $this->cpf->validaCPF($input['cpf']);
 
-        $deletedRows = Cpf::where('cpf', '=',  $input['cpf'])->delete();
+            if($valida_cpf)
+            {
+                $deletedRows = Cpf::where('cpf', '=',  $input['cpf'])->delete();
 
-        return $this->sendResponse($deletedRows, 'Cpf deletado.');
+                return $this->sendResponse($deletedRows, 'Cpf deletado.')->setStatusCode(200);
+            }
+
+            return $this->sendError('cpf invalido.')->setStatusCode(404);
+
+        } catch (Exception $e) {
+            return $this->sendResponse($input['cpf'], $e->getMessage)->setStatusCode(400);; 
+        }
     }
 
     /**
@@ -116,12 +236,12 @@ class ConsultaController extends BaseController
 
         $listagem['blacklist'] =  Cpf::where('blocked', '=', 1)->count();
 
-        if(!isset($_SESSION['postCounter']))
+        if(!isset($_SESSION['consultas_counter']))
         {
-            $_SESSION['postCounter'] = 0;
+            $_SESSION['consultas_counter'] = 0;
         }
         
-        $listagem['consultas_realizadas'] = $_SESSION['postCounter'];
+        $listagem['consultas_realizadas'] = $_SESSION['consultas_counter'];
         //Cpf::sum('count');
 
         $to = \Carbon\Carbon::createFromFormat('Y-m-d H:i:s', $_SESSION['start_date']);
@@ -129,6 +249,6 @@ class ConsultaController extends BaseController
 
         $listagem['uptime'] = $to->diffInMinutes($from).' minutos';
 
-        return $this->sendResponse($listagem, 'Status serviço');
+        return $this->sendResponse($listagem, 'Status serviço')->setStatusCode(200);
     }
 }
